backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
pydantic[email]
bcrypt==4.0.1 
passlib==1.7.4